fix(app.spec): fixed store configuration
